chore: move grpc and utility logic to internal
Move the grpc code and the utility code that customers aren't expected
to interact with into the internal path. This reduces the logic that
users have to grok when using the sdk.
