+++ integration/list.test.ts
@@ -1,5 +1,5 @@
 import {v4} from 'uuid';
-import {sleep} from '../src/utils/sleep';
+import {sleep} from '../src/internal/utils/sleep';
 import {
   CollectionTtl,
   CacheListConcatenateBack,
