+++ src/internal/utils/collections.ts
