+++ src/internal/utils/display.ts
