+++ src/messages/responses/cache-list-fetch.ts
@@ -6,7 +6,7 @@ import {
   ResponseMiss,
 } from './response-base';
 import {TextDecoder} from 'util';
-import {truncateStringArray} from '../../utils/display';
+import {truncateStringArray} from '../../internal/utils/display';
 
 const TEXT_DECODER = new TextDecoder();
