+++ src/messages/responses/cache-set-fetch.ts
@@ -6,7 +6,7 @@ import {
 } from './response-base';
 import {SdkError} from '../../errors/errors';
 import {TextDecoder} from 'util';
-import {truncateStringArray} from '../../utils/display';
+import {truncateStringArray} from '../../internal/utils/display';
 
 const TEXT_DECODER = new TextDecoder();
